Add variable to allow fonts-config to update default settings
This variable tells the %post scriptlet of the fonts-config
package if it should update the settings in
/etc/sysconfig/fonts-config with the default values. The
FORCE_MODIFY_DEFAULT_FONT_SETTINGS_IN_NEXT_UPDATE variable would
rarely be set to "yes" but the fact that the value doesn't exist
in a system would already tell the package the default settings
should be changed once.

This is done this way because fillup doesn't modify existing values
from the changed templates, but we want users to have the latest
default values.

This (and the associated change in the %post scriptlet in
fonts-config) should fix boo#1172022.